refactor: Move ResponseError down to rest package
