+++ pkg/client/dtclient/client.go
@@ -27,7 +27,6 @@ import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/version"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/api"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client/auth"
-	clientErrors "github.com/dynatrace/dynatrace-configuration-as-code/pkg/client/errors"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client/metadata"
 	dtVersion "github.com/dynatrace/dynatrace-configuration-as-code/pkg/client/version"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2/coordinate"
@@ -458,16 +457,16 @@ func (d *DynatraceClient) upsertSettings(ctx context.Context, obj SettingsObject
 
 	resp, err := rest.SendWithRetryWithInitialTry(ctx, d.client, rest.Post, obj.Coordinate.ConfigId, requestUrl, payload, d.retrySettings.Normal)
 	if err != nil {
-		return DynatraceEntity{}, clientErrors.NewRespErr("failed to create or update Settings object", resp).WithErr(err)
+		return DynatraceEntity{}, rest.NewRespErr("failed to create or update Settings object", resp).WithErr(err)
 	}
 
 	if !resp.IsSuccess() {
-		return DynatraceEntity{}, clientErrors.NewRespErr(fmt.Sprintf("failed to create or update Settings object with externalId %s (HTTP %d)!\n\tResponse was: %s", externalID, resp.StatusCode, string(resp.Body)), resp)
+		return DynatraceEntity{}, rest.NewRespErr(fmt.Sprintf("failed to create or update Settings object with externalId %s (HTTP %d)!\n\tResponse was: %s", externalID, resp.StatusCode, string(resp.Body)), resp)
 	}
 
 	entity, err := parsePostResponse(resp)
 	if err != nil {
-		return DynatraceEntity{}, clientErrors.NewRespErr("failed to parse response", resp).WithErr(err)
+		return DynatraceEntity{}, rest.NewRespErr("failed to parse response", resp).WithErr(err)
 	}
 
 	log.Debug("\tCreated/Updated object %s (%s) with externalId %s", obj.Coordinate.ConfigId, obj.SchemaId, externalID)
@@ -598,13 +597,13 @@ func (d *DynatraceClient) listSchemas() (SchemaList, error) {
 	}
 
 	if !resp.IsSuccess() {
-		return nil, clientErrors.NewRespErr(fmt.Sprintf("request failed with HTTP (%d).\n\tResponse content: %s", resp.StatusCode, string(resp.Body)), resp)
+		return nil, rest.NewRespErr(fmt.Sprintf("request failed with HTTP (%d).\n\tResponse content: %s", resp.StatusCode, string(resp.Body)), resp)
 	}
 
 	var result SchemaListResponse
 	err = json.Unmarshal(resp.Body, &result)
 	if err != nil {
-		return nil, clientErrors.NewRespErr("failed to unmarshal response", resp).WithErr(err)
+		return nil, rest.NewRespErr("failed to unmarshal response", resp).WithErr(err)
 	}
 
 	if result.TotalCount != len(result.Items) {
@@ -636,14 +635,14 @@ func (d *DynatraceClient) getSettingById(objectId string) (*DownloadSettingsObje
 		// special case of settings API: If you give it any object ID for a setting object that does not exist, it will give back 400 BadRequest instead
 		// of 404 Not found. So we interpret both status codes, 400 and 404, as "not found"
 		if resp.StatusCode == http.StatusBadRequest || resp.StatusCode == http.StatusNotFound {
-			return nil, clientErrors.NewRespErr(ErrSettingNotFound.Error(), resp).WithErr(ErrSettingNotFound)
+			return nil, rest.NewRespErr(ErrSettingNotFound.Error(), resp).WithErr(ErrSettingNotFound)
 		}
-		return nil, clientErrors.NewRespErr(fmt.Sprintf("request failed with HTTP (%d).\n\tResponse content: %s", resp.StatusCode, string(resp.Body)), resp)
+		return nil, rest.NewRespErr(fmt.Sprintf("request failed with HTTP (%d).\n\tResponse content: %s", resp.StatusCode, string(resp.Body)), resp)
 	}
 
 	var result DownloadSettingsObject
 	if err = json.Unmarshal(resp.Body, &result); err != nil {
-		return nil, clientErrors.NewRespErr("failed to unmarshal response", resp).WithErr(err)
+		return nil, rest.NewRespErr("failed to unmarshal response", resp).WithErr(err)
 	}
 
 	return &result, nil
@@ -693,7 +692,7 @@ func (d *DynatraceClient) listSettings(schemaId string, opts ListSettingsOptions
 
 	u, err := buildUrl(d.environmentURL, d.settingsObjectAPIPath, params)
 	if err != nil {
-		return nil, clientErrors.RespError{Err: err}
+		return nil, rest.RespError{Err: err}
 	}
 
 	_, err = rest.ListPaginated(d.client, d.retrySettings, u, schemaId, addToResult)
@@ -748,7 +747,7 @@ func (d *DynatraceClient) listEntitiesTypes() ([]EntitiesType, error) {
 
 	u, err := buildUrl(d.environmentURL, pathEntitiesTypes, params)
 	if err != nil {
-		return nil, clientErrors.RespError{Err: err}
+		return nil, rest.RespError{Err: err}
 	}
 
 	_, err = rest.ListPaginated(d.client, d.retrySettings, u, "EntityTypeList", addToResult)
@@ -807,7 +806,7 @@ func (d *DynatraceClient) listEntities(entitiesType EntitiesType) ([]string, err
 
 		u, err := buildUrl(d.environmentURL, pathEntitiesObjects, params)
 		if err != nil {
-			return nil, clientErrors.RespError{Err: err}
+			return nil, rest.RespError{Err: err}
 		}
 
 		resp, err := rest.ListPaginated(d.client, d.retrySettings, u, entityType, addToResult)
