+++ pkg/rest/retry.go
@@ -19,7 +19,6 @@ package rest
 import (
 	"context"
 	"fmt"
-	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client/errors"
 	"net/http"
 	"time"
 
@@ -74,8 +73,8 @@ func GetWithRetry(client *http.Client, url string, settings RetrySetting) (resp 
 		return resp, fmt.Errorf("GET request %s failed after %d retries: %w", url, settings.MaxRetries, err)
 	}
 
-	return resp, errors.RespError{
-		Type:       errors.RespErrType,
+	return resp, RespError{
+		Type:       RespErrType,
 		StatusCode: resp.StatusCode,
 		Message:    fmt.Sprintf("GET request %s failed after %d retries: (HTTP %d)!\n    Response was: %s", url, settings.MaxRetries, resp.StatusCode, resp.Body),
 		Body:       string(resp.Body),
@@ -95,9 +94,9 @@ func SendWithRetry(ctx context.Context, client *http.Client, sendWithBody SendRe
 	}
 
 	if err != nil {
-		return Response{}, errors.RespError{Type: errors.RespErrType, StatusCode: resp.StatusCode, Message: err.Error()}
+		return Response{}, fmt.Errorf("HTTP send request %s failed after %d retries: %w", path, setting.MaxRetries, err)
 	}
-	return Response{}, errors.RespError{Type: errors.RespErrType, StatusCode: resp.StatusCode, Body: string(resp.Body)}
+	return Response{}, NewRespErr(fmt.Sprintf("HTTP send request %s failed after %d retries: (HTTP %d)!\n    Response was: %s", path, setting.MaxRetries, resp.StatusCode, string(resp.Body)), resp)
 }
 
 // SendWithRetryWithInitialTry will try to call sendWithBody and if it didn't succeed call [SendWithRetry]
